refactor(core): Extract HTTP error utilities to reduce file size
Split HTTP-specific error creation logic from transport-error-factory.ts
to transport-error-http-utils.ts to resolve max-lines ESLint violation.

Changes:
- Created transport-error-http-utils.ts with:
  - isHttpStatusRetryable() helper function
  - createErrorFromHttpStatus() factory function
- Updated transport-error-factory.ts to re-export HTTP utilities
- Reduced main factory file from 441 lines to 356 lines (under 400 limit)
- Maintained backward compatibility via re-export
- No changes to public API
